server/requirements.txt
flask==1.0.2
flask-cors==3.0.7
pymongo==3.11.0
Flask-PyMongo==2.3.0
dnspython
gunicorn>=20.0.4
numpy
requests
python-dotenv
matplotlib